intranet/scheduler/requirements.txt
grpcio
grpcio-tools

fastapi
uvicorn[standard]

apscheduler

pydantic-settings